+++ dart/geodata/lib/formats.dart
@@ -1,0 +1,13 @@
+// Copyright (c) 2020-2023 Navibyte (https://navibyte.com). All rights reserved.
+// Use of this source code is governed by a “BSD-3-Clause”-style license that is
+// specified in the LICENSE file.
+//
+// Docs: https://github.com/navibyte/geospatial
+
+/// OpenAPI document and Common Query Language (CQL2) formats (partial support).
+///
+/// Usage: import `package:geodata/formats.dart`
+library formats;
+
+export 'src/formats/cql2/cql_query.dart';
+export 'src/formats/openapi/open_api_document.dart';