+++ dart/geodata/lib/geodata.dart
@@ -12,5 +12,6 @@ library geodata;
 // Export mini-libraries forming the whole "geodata" library.
 export 'common.dart';
 export 'core.dart';
+export 'formats.dart';
 export 'geojson_client.dart';
 export 'ogcapi_features_client.dart';