+++ dart/geodata/lib/src/ogcapi_common/service/client/ogc_client.dart
@@ -9,7 +9,7 @@ import 'package:meta/meta.dart';
 import '/src/common/links/link.dart';
 import '/src/common/links/links.dart';
 import '/src/common/service/service_exception.dart';
-import '/src/core/api/open_api_document.dart';
+import '/src/formats/openapi/open_api_document.dart';
 import '/src/ogcapi_common/model/ogc_service.dart';
 import '/src/ogcapi_common/model/ogc_service_meta.dart';
 import '/src/utils/cached_object.dart';